move test to SQLQuerySuite
